\name{ISfit-package}
\alias{ISfit-package}
\alias{ISfit}
\docType{package}
\title{
Incoherent scatter plasma parameter fits
}
\description{
Incoherent scatter plasma parameter estimation for multistatic radar systems
}
\details{
\tabular{ll}{
Package: \tab ISfit\cr
Version: \tab 0.1-5\cr
Date: \tab 2014-01-27\cr
Depends: \tab IGRF11, ISspectrum, NLSQ, IRI2011, radarPointings, FME, lattice\cr
}

  Index:

  \code{\link{plotPlasmaParams}} Plot plasma parameters as function of height (and time)

  \code{\link{plotMarginalDistribution}} Plot marginal distributions of plasma parameters using results of MCMC analysis

}

To be completed, please be patient...

\author{
Ilkka Virtanen (University of Oulu, Finland) \cr \email{ilkka.i.virtanen@oulu.fi}
}
\keyword{ package }
