\name{PPIgraphics}
\alias{ppig}
\alias{ppigraphics}

\title{PPIgraphics}

\description{Graphics package for Plasma Parameter Inversion}

\details{PPIgraphics contains routines for plotting Plasma Parameter Inversion results.

\bold{PPIgraphics functions:}

\code{\link{plotPlasmaParams}} Plot plasma parameters as function of height (and time)

\code{\link{plotMarginalDistribution}} Plot marginal distributions of plasma parameters using results of MCMC analysis

}



\note{ TO BE COMPLETED, please be patient.. :) }

\author{Ilkka Virtanen (University of Oulu, Finland) \email{ilkka.i.virtanen@oulu.fi}}
