\name{ISfit-package}
\alias{ISfit-package}
\alias{ISfit}
\docType{package}
\title{
Incoherent scatter plasma parameter fits
}
\description{
Incoherent scatter plasma parameter estimation for multistatic radar systems
}
\details{
\tabular{ll}{
Package: \tab ISfit\cr
Version: \tab 0.1-0\cr
Date: \tab 2012-01-13\cr
License: \tab GPL-2\cr
Depends: \tab IGRF11, ISspectrum, NLSQ, iri2011\cr
Built: \tab R 2.14.1; ; 2012-01-13 07:47:55 UTC; unix\cr
}
}
\author{
Ilkka Virtanen (University of Oulu, Finland) \cr \email{ilkka.i.virtanen@oulu.fi}
}
\keyword{ package }
